# autogened-swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "RUGENDO API",
    "description": "Made by Igor Roggy"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "petstore_auth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/mtn-token": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/requestToPay": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "x-reference-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/mtn-payment-status/{referenceId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "referenceId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/send-to-email": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "header",
            "type": "string"
          },
          {
            "name": "subject",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "req.headersData": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}